+++ R/detectSerieOut.R
@@ -7,7 +7,10 @@
 #' correlation to the majority of time courses. To support the analysis by
 #' correlations, a principal component analysis is done on the plant
 #' (time course) by spline coefficient matrix. A PCA plot of the plant scores
-#' will show the outlying plants.
+#' will show the outlying plants. Finally the pairwise-ratios of the slopes of
+#' a linear model fitted through the spline coefficients are computed. Plants
+#' are tagged when the average pairwise-ratio is lower the a given threshold
+#' (`thrSlope`).
 #'
 #' @param corrDat A data.frame with corrected spatial data.
 #' @param predDat A data.frame with predicted data from a fitted spline.
@@ -23,6 +26,8 @@
 #' based on correlation between plots.
 #' @param thrPca A numerical value used as threshold for determining outliers
 #' based on angles (in degrees) between PCA scores.
+#' @param thrSlope A numerical value used as threshold for determining outliers
+#' based on slopes.
 #'
 #' @return An object of class \code{serieOut}, a \code{data.frame} with outlying
 #' series of observations.
@@ -50,7 +55,8 @@
 #'                            trait = "EffpsII_corr",
 #'                            genotypes = subGenoVator,
 #'                            thrCor = 0.9,
-#'                            thrPca = 30)
+#'                            thrPca = 30,
+#'                            thrSlope = 0.7)
 #'
 #' ## The `outVator` can be visualized for selected genotypes.
 #' plot(outVator, genotypes = "G151")
@@ -67,7 +73,8 @@ detectSerieOut <- function(corrDat,
                            genotypes = NULL,
                            geno.decomp = NULL,
                            thrCor = 0.9,
-                           thrPca = 30) {
+                           thrPca = 30,
+                           thrSlope = 0.7) {
   ## Checks.
   if (!is.character(trait) || length(trait) > 1) {
     stop("trait should be a character string of length 1.\n")
@@ -112,11 +119,14 @@ detectSerieOut <- function(corrDat,
   if (!all(genotypes %in% corrDat[["genotype"]])) {
     stop("all genotypes should be in corrDat")
   }
-  if (!is.numeric(thrCor) || any(thrCor < 0) || any(thrCor > 1)) {
-    stop("thrCor should be a numerical vector with values between 0 and 1.\n")
+  if (!is.numeric(thrCor) || any(thrCor < -1) || any(thrCor > 1)) {
+    stop("thrCor should be a numerical vector with values between -1 and 1.\n")
   }
   if (!is.numeric(thrPca) || any(thrPca < 0)) {
     stop("thrPca should be a numerical vector with positive values.\n")
+  }
+  if (!is.numeric(thrSlope) || any(thrSlope < 0) || any(thrSlope > 1)) {
+    stop("thrSlope should be a numerical vector with values between 0 and 1.\n")
   }
   genoPlotId <- sapply(X = genotypes, FUN = function(genotype) {
     length(unique(corrDat[corrDat[["genotype"]] == genotype, "plotId"]))
@@ -151,12 +161,22 @@ detectSerieOut <- function(corrDat,
       stop("thrPca should be a named vector, with names matching the levels ",
            "in geno.decomp.\n")
     }
+    if (length(thrSlope) == 1) {
+      thrSlope <- setNames(rep(thrSlope, times = nGenoDecomp), genoDecompLevs)
+    } else if (is.null(names(thrSlope)) ||
+               !all(genoDecompLevs %in% names(thrSlope))) {
+      stop("thrSlope should be a named vector, with names matching the levels ",
+           "in geno.decomp.\n")
+    }
   } else {
     if (length(thrCor) != 1) {
       stop("thrCor should be a vector of length 1.\n")
     }
     if (length(thrPca) != 1) {
       stop("thrPca should be a vector of length 1.\n")
+    }
+    if (length(thrSlope) != 1) {
+      stop("thrSlope should be a vector of length 1.\n")
     }
   }
   ## Restrict predDat and coefDat to genotypes.
@@ -195,13 +215,13 @@ detectSerieOut <- function(corrDat,
                                                     formula = type ~ plotId,
                                                     value.var = "obj.coefficients")
                         ## Remove intercept.
-                        plantDat <- plantDat[-1, ]
+                        plantDat <- plantDat[-1, , drop = FALSE]
                         ## Remove plants with only NA.
                         NAplants <- apply(X = plantDat, MARGIN = 2,
                                           FUN = function(plant) {
                                             all(is.na(plant))
                                           })
-                        plantDat <- plantDat[, !NAplants]
+                        plantDat <- plantDat[, !NAplants, drop = FALSE]
                         ## Add geno.decomp as attribute for later use.
                         if (!is.null(geno.decomp)) {
                           attr(plantDat, which = "genoDecomp") <-
@@ -209,7 +229,7 @@ detectSerieOut <- function(corrDat,
                         }
                         return(plantDat)
                       })
-  ## Compute correlation matrix.
+  ## Compute correlation matrices.
   cormats <- lapply(X = plantDats, FUN = function(plantDat) {
     if (!is.null(dim(plantDat))) {
       ## if there are plants, estimate the correlation...
@@ -241,6 +261,41 @@ detectSerieOut <- function(corrDat,
       return(NULL)
     }
   })
+  ## Compute slope matrices.
+  slopemats <- lapply(X = plantDats, FUN = function(plantDat) {
+    if (!is.null(dim(plantDat))) {
+      ## if there are plants, estimate the slopes.
+      ## Convert matrix to data.frame for lm.
+      plantDat <- as.data.frame(plantDat)
+      ## Construct empty matrix for storing results.
+      slopemat <- matrix(nrow = ncol(plantDat), ncol = ncol(plantDat),
+                       dimnames = list(colnames(plantDat), colnames(plantDat)))
+      ## Compute slope per pair of plots.
+      slopemat[lower.tri(slopemat)] <-
+        combn(x = colnames(plantDat), m = 2, FUN = function(plants) {
+          ## Fit linear model and extract slope.
+          modForm <- formula(paste(plants, collapse = "~"))
+          slope <- abs(coef(lm(modForm, data = plantDat))[2])
+          if (slope > 1) slope <- 1 / slope
+          return(slope)
+        }, simplify = TRUE)
+      ## Fill upper triangle of slopemat matrix with copy of lower triangle.
+      slopemat[upper.tri(slopemat)] <- t(slopemat)[upper.tri(slopemat)]
+      attr(slopemat, which = "genoDecomp") <-
+        attr(plantDat, which = "genoDecomp")
+      if (!is.null(geno.decomp)) {
+        attr(slopemat, which = "thrSlope") <-
+          thrSlope[attr(plantDat, which = "genoDecomp")]
+      } else {
+        attr(slopemat, which = "thrSlope") <- thrSlope
+      }
+      return(slopemat)
+    } else {
+      ## ... if not, return a null matrix
+      return(NULL)
+    }
+  })
+  ## Find outlying plots based on correlation.
   annotatePlantsCor <- lapply(X = names(cormats), FUN = function(geno) {
     if (!is.null(cormats[[geno]])) {
       ## Compute mean based on all but the worst correlation per plant.
@@ -262,6 +317,7 @@ detectSerieOut <- function(corrDat,
       return(NULL)
     }
   })
+  ## Convert cormats to format used by ggplot.
   cormats <- lapply(X = cormats, FUN = function(cormat) {
     ## Set lower part of cormat to NA for plotting.
     cormat[lower.tri(cormat)] <- NA
@@ -270,6 +326,7 @@ detectSerieOut <- function(corrDat,
     attr(meltedCormat, which = "thrCor") <- attr(cormat, which = "thrCor")
     return(meltedCormat)
   })
+  ## Find outlying plots based on PCA angles.
   annotatePlantsPcaAngle <- lapply(X = names(plantPcas), FUN = function(geno) {
     ## Calculate the pairwise difference of coordinates on the 2nd axis and
     ## annotate plant with average diff larger than threshold.
@@ -289,7 +346,7 @@ detectSerieOut <- function(corrDat,
       thrPcaPlant <- thrPca
     }
     if (any(meanPcAngles >= thrPcaPlant)) {
-      ## Create data.frame with info on plants with average difference.
+      ## Create data.frame with info on plants with average difference
       ## above threshold.
       annPlotsPcAngle <- meanPcAngles[meanPcAngles >= thrPcaPlant]
       return(data.frame(plotId = names(annPlotsPcAngle), reason = "angle",
@@ -298,8 +355,42 @@ detectSerieOut <- function(corrDat,
       return(NULL)
     }
   })
+  ## Find outlying plots based on slopes.
+  annotatePlantsSlope <- lapply(X = names(slopemats), FUN = function(geno) {
+    if (!is.null(slopemats[[geno]])) {
+      ## Compute mean slope per plot on all but the worst slope per plant.
+      ## This prevents one bad correlation causing all plants to be outliers.
+      meanSlope <- apply(slopemats[[geno]], MARGIN = 1,
+                         FUN = function(plant) {
+                           mean(sort(plant)[-1])
+                         })
+      thrSlopePlant <- attr(slopemats[[geno]], which = "thrSlope")
+    } else {
+      meanSlope <- NULL
+    }
+    if (any(meanSlope < thrSlopePlant)) {
+      ## Create data.frame with info on plants with average difference
+      ## above threshold.
+      annPlotsSlope <- meanSlope[meanSlope < thrSlopePlant]
+      return(data.frame(plotId = names(annPlotsSlope), reason = "slope",
+                        value = annPlotsSlope))
+    } else {
+      return(NULL)
+    }
+  })
+  ## Convert slopemats to format used by ggplot.
+  slopemats <- lapply(X = slopemats, FUN = function(slopemat) {
+    ## Set lower part of cormat to NA for plotting.
+    slopemat[lower.tri(slopemat)] <- NA
+    ## Melt to format used by ggplot.
+    meltedSlopemat <- reshape2::melt(slopemat, na.rm = TRUE)
+    attr(meltedSlopemat, which = "thrCor") <- attr(slopemat, which = "thrCor")
+    return(meltedSlopemat)
+  })
   ## Create full data.frame with annotated plants.
-  annotatePlants <- do.call(rbind, c(annotatePlantsCor, annotatePlantsPcaAngle))
+  annotatePlants <- do.call(rbind, c(annotatePlantsCor,
+                                     annotatePlantsPcaAngle,
+                                     annotatePlantsSlope))
   if (!is.null(annotatePlants)) {
     ## Merge genotype and geno.decomp to annotated plants.
     annotatePlants <- merge(unique(corrDatPred[c("genotype", geno.decomp,
@@ -320,11 +411,13 @@ detectSerieOut <- function(corrDat,
   class(annotatePlants) <- c("serieOut", class(annotatePlants))
   attr(x = annotatePlants, which = "thrCor") <- thrCor
   attr(x = annotatePlants, which = "thrPca") <- thrPca
+  attr(x = annotatePlants, which = "thrSlope") <- thrSlope
   attr(x = annotatePlants, which = "trait") <- trait
   attr(x = annotatePlants, which = "geno.decomp") <- geno.decomp
   attr(x = annotatePlants, which = "plotInfo") <- plotInfo
   attr(x = annotatePlants, which = "cormats") <- cormats
   attr(x = annotatePlants, which = "plantPcas") <- plantPcas
+  attr(x = annotatePlants, which = "slopemats") <- slopemats
   attr(x = annotatePlants, which = "genoPreds") <- genoPreds
   attr(x = annotatePlants, which = "genoDats") <- genoDats
   return(annotatePlants)
@@ -341,7 +434,10 @@ detectSerieOut <- function(corrDat,
 #' @param x An object of class \code{serieOut}.
 #' @param ... Ignored.
 #' @param genotypes A character vector indicating which genotypes should be
-#' plotted.
+#' plotted. If \code{NULL} all genotypes are plotted.
+#' @param geno.decomp A character vector indicating which levels of
+#' \code{geno.decomp} should be plotted. If \code{NULL} all levels are plotted.
+#' Ignored if \code{geno.decomp} was not used when fitting models.
 #' @param useTimeNumber Should the timeNumber be used instead of the timePoint
 #' in the labels on the x-axis?
 #' @param timeNumber If \code{useTimeNumber = TRUE}, a character vector
@@ -372,7 +468,8 @@ detectSerieOut <- function(corrDat,
 #'                            trait = "EffpsII_corr",
 #'                            genotypes = subGenoVator,
 #'                            thrCor = 0.9,
-#'                            thrPca = 30)
+#'                            thrPca = 30,
+#'                            thrSlope = 0.7)
 #'
 #' ## The `outVator` can be visualized for selected genotypes.
 #' plot(outVator, genotypes = "G151")
@@ -384,6 +481,7 @@ detectSerieOut <- function(corrDat,
 plot.serieOut <- function(x,
                           ...,
                           genotypes = NULL,
+                          geno.decomp = NULL,
                           useTimeNumber = FALSE,
                           timeNumber = NULL,
                           title = NULL,
@@ -395,26 +493,36 @@ plot.serieOut <- function(x,
   thrCor <- attr(x = x, which = "thrCor")
   thrPca <- attr(x = x, which = "thrPca")
   trait <- attr(x = x, which = "trait")
-  geno.decomp <- attr(x = x, which = "geno.decomp")
+  geno.decompVar <- attr(x = x, which = "geno.decomp")
   plotInfo <- attr(x = x, which = "plotInfo")
+  ## Restrict to selected levels of geno.decomp.
+  if (!is.null(geno.decomp) && !is.null(geno.decompVar)) {
+    if (!all(geno.decomp %in% plotInfo[[geno.decompVar]])) {
+      stop("All selected geno.decomp levels should be in the data.\n")
+    }
+    plotInfo <- plotInfo[plotInfo[[geno.decompVar]] %in% geno.decomp, ]
+  }
   if (!is.null(genotypes) && (!is.character(genotypes) ||
                               !all(genotypes %in% plotInfo[["genotype"]]))) {
     stop("genotypes should be a character vector of genotypes used for ",
          "outlier detection.\n")
   }
   if (is.null(genotypes)) {
-    genotypes <- interaction(plotInfo[c("genotype", geno.decomp)])
+    genotypes <- interaction(plotInfo[c("genotype", geno.decompVar)])
   } else {
     genotypes <- interaction(plotInfo[plotInfo[["genotype"]] %in% genotypes,
-                                      c("genotype", geno.decomp)])
+                                      c("genotype", geno.decompVar)])
   }
   genotypes <- as.character(genotypes)
   cormats <- attr(x = x, which = "cormats")[genotypes]
   plantPcas <- attr(x = x, which = "plantPcas")[genotypes]
+  slopemats <- attr(x = x, which = "slopemats")[genotypes]
   genoPreds <- attr(x = x, which = "genoPreds")[genotypes]
   genoDats <- attr(x = x, which = "genoDats")[genotypes]
   ## Get minimum correlation. Cannot be higher than 0.8.
   minCor <- min(c(unlist(cormats, use.names = FALSE), 0.8), na.rm = TRUE)
+  ## Get minimum slope. Cannot be higher than 0.8.
+  minSlope <- min(c(unlist(slopemats, use.names = FALSE), 0.8), na.rm = TRUE)
   ## Compute the number of breaks for the time scale based on all plants.
   ## If there are less than 3 time points use the number of time points.
   ## Otherwise use 3.
@@ -460,6 +568,7 @@ plot.serieOut <- function(x,
     }
     ## Correlation plot.
     thrCorGeno <- attr(cormats[[genotype]], which = "thrCor")
+    thrSlopeGeno <- attr(cormats[[genotype]], which = "thrSlope")
     correl <- ggplot2::ggplot(data = cormats[[genotype]],
                               ggplot2::aes_string("Var2", "Var1",
                                                   fill = "value")) +
@@ -469,6 +578,16 @@ plot.serieOut <- function(x,
                                                                thrCorGeno, 1)),
                                     limits = c(minCor, 1),
                                     name = "Pearson\nCorrelation") +
+      ggnewscale::new_scale_fill() +
+      ## Add slope to upper left.
+      ggplot2::geom_tile(data = slopemats[[genotype]],
+                         ggplot2::aes_string("Var1", "Var2", fill = "value"),
+                         color = "white") +
+      ggplot2::scale_fill_gradientn(colors = c("cyan", "white", "darkgreen"),
+                                     values = scales::rescale(c(minSlope,
+                                                                thrSlopeGeno, 1)),
+                                     limits = c(minSlope, 1),
+                                     name = "Slope\nCorrelation") +
       ## Move y-axis to the right.
       ggplot2::scale_y_discrete(position = "right") +
       ## Use coord fixed to create a square shaped output.
@@ -478,10 +597,13 @@ plot.serieOut <- function(x,
                      panel.grid = ggplot2::element_line(color = "grey92"),
                      plot.title = ggplot2::element_text(hjust = 0.5),
                      axis.ticks = ggplot2::element_blank(),
+                     panel.grid.major = ggplot2::element_blank(),
+                     panel.grid.minor = ggplot2::element_blank(),
                      axis.text.x = ggplot2::element_text(angle = 45, vjust = 1,
-                                                         size = 12, hjust = 1),
-                     legend.position = "left") +
-      ggplot2::labs(title = "Correl of coef", x = NULL, y = NULL)
+                                                         hjust = 1),
+                     legend.position = "left",
+                     legend.box = "horizontal") +
+      ggplot2::labs(title = "Correlations", x = NULL, y = NULL)
     ## PCA biplot.
     pcaplot <- factoextra::fviz_pca_var(plantPcas[[genotype]])
     ## Arrange plots.
@@ -522,15 +644,16 @@ plot.serieOut <- function(x,
 #' \code{\link{fitSpline}} function.
 #' @param serieOut A data.frame with at least the column plotId with
 #' values corresponding to those in dat/fitSpline.
-#' @param trait The trait that should be replaced by NA. Can be ignored when
-#' using the output of \code{detectSerieOut} as input.
+#' @param traits The traits that should be replaced by NA. When using the
+#' output of \code{detectSerieOut} as input for \code{serieOut} this defaults
+#' to the trait used for when detecting the outliers.
 #'
 #' @return Depending on the input either a \code{data.frame} or an object of
 #' class \code{HTPSpline} for which the outliers specified in \code{serieOut}
-#' are replace by NA.
+#' are replaced by NA.
 #'
 #' @examples
-#' ## Run the function to fit P-spline on a subset of genotypes.
+#' ## Run the function to fit P-splines on a subset of genotypes.
 #' subGenoVator <- c("G160", "G151")
 #' fit.spline <- fitSpline(inDat = spatCorrectedVator,
 #'                         trait = "EffpsII_corr",
@@ -548,11 +671,19 @@ plot.serieOut <- function(x,
 #'                            trait = "EffpsII_corr",
 #'                            genotypes = subGenoVator,
 #'                            thrCor = 0.9,
-#'                            thrPca = 30)
+#'                            thrPca = 30,
+#'                            thrSlope = 0.7)
 #'
 #' ## Replace the outliers by NA in the corrected data.
 #' spatCorrectedVatorOut <- removeSerieOut(dat = spatCorrectedVator,
 #'                                         serieOut = outVator)
+#'
+#' ## Replace the outliers by NA in the corrected data.
+#' ## Replace both the corrected value and the raw trait value by NA.
+#' spatCorrectedVatorOut2 <-
+#'   removeSerieOut(dat = spatCorrectedVator,
+#'                  serieOut = outVator,
+#'                  traits = c("EffpsII", "EffpsII_corr"))
 #'
 #' @family functions for detecting outliers for series of observations
 #'
@@ -560,7 +691,7 @@ plot.serieOut <- function(x,
 removeSerieOut <- function(dat = NULL,
                            fitSpline = NULL,
                            serieOut,
-                           trait = attr(x = serieOut,
+                           traits = attr(x = serieOut,
                                         which = "trait")) {
   ## Check that one of dat and fitSpline are specified.
   if ((is.null(dat) && is.null(fitSpline)) || (
@@ -577,7 +708,7 @@ removeSerieOut <- function(dat = NULL,
   }
   if (!is.null(fitSpline)) {
     if (!inherits(fitSpline, "HTPSpline")) {
-      stop("dat should be an object of class HTPSpline.\n")
+      stop("fitSpline should be an object of class HTPSpline.\n")
     }
   }
   if (!inherits(serieOut, "data.frame")) {
@@ -589,14 +720,18 @@ removeSerieOut <- function(dat = NULL,
     }
     if (!is.null(dat)) {
       ## Remove plots that are in serieOut.
-      dat[dat[["plotId"]] %in% serieOut[["plotId"]], trait] <- NA
+      for (trait in traits) {
+        dat[dat[["plotId"]] %in% serieOut[["plotId"]], trait] <- NA
+      }
     } else if (!is.null(fitSpline)) {
       fitSpline$coefDat[fitSpline$coefDat[["plotId"]] %in%
                           serieOut[["plotId"]], "obj.coefficients"] <- NA
-      fitSpline$predDat[fitSpline$predDat[["plotId"]] %in%
-                          serieOut[["plotId"]], c("pred.value", "deriv")] <- NA
+      fitSpline$predDat[fitSpline$predDat[["plotId"]] %in% serieOut[["plotId"]],
+                        c("pred.value", "deriv", "deriv2")] <- NA
       modDat <- attr(x = fitSpline, which = "modDat")
-      modDat[modDat[["plotId"]] %in% serieOut[["plotId"]], trait] <- NA
+      for (trait in traits) {
+        modDat[modDat[["plotId"]] %in% serieOut[["plotId"]], trait] <- NA
+      }
       attr(x = fitSpline, which = "modDat") <- modDat
     }
   }